project/requirements/base.txt
Pillow==6.0.0
Django==2.2.2
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
django-split-settings==0.3.0
drf-yasg==1.15.0
psycopg2-binary==2.8.3
